Handling review comments related to code formatting
Changed 'uint8_t ht_idx' to 'uint16_t ht_idx' in beck.c

+++ be/tool/beck.c
@@ -1250,14 +1250,15 @@ static struct m0_stob_ad_domain *emap_dom_find(const struct action *act,
 {
 	struct m0_stob_ad_domain *adom = NULL;
 	int			  i;
-	uint8_t			  ht_idx;
+	uint16_t		  ht_idx;
 
 	ht_idx = m0_stob_get_hash(emap_fid);
 
 	for (i = 0; i < act->a_builder->b_ad_dom_count; i++) {
 		adom = act->a_builder->b_ad_domain[i];
 		if (m0_fid_eq(emap_fid,
-	            &adom->sad_adata_ht[ht_idx].sad_adata.em_mapping.bb_backlink.bli_fid)) {
+	            &adom->sad_adata_ht[ht_idx].sad_adata.em_mapping.
+                          bb_backlink.bli_fid)) {
 			break;
 		}
 	}
@@ -1356,15 +1357,15 @@ static int emap_entry_lookup(struct m0_stob_ad_domain  *adom,
 {
 	int                       rc;
 	struct m0_fid             fid;
-	uint8_t                   ht_idx;
+	uint16_t                  ht_idx;
 
 	fid.f_container = prefix.u_hi;
 	fid.f_key = prefix.u_lo;
 	ht_idx = m0_stob_get_hash(&fid);
 	M0_LOG(M0_DEBUG, U128X_F, U128_P(&prefix));
 	rc = M0_BE_OP_SYNC_RET_WITH( &it->ec_op,
-				  m0_be_emap_lookup(&adom->sad_adata_ht[ht_idx].sad_adata,
-						    &prefix, offset, it),
+	                m0_be_emap_lookup(&adom->sad_adata_ht[ht_idx].sad_adata,
+                                         &prefix, offset, it),
 				  bo_u.u_emap.e_rc);
 	return rc == -ESRCH ? -ENOENT : rc;
 }
@@ -1379,7 +1380,7 @@ static int emap_prep(struct action *act, struct m0_be_tx_credit *credit)
 	struct m0_be_emap_cursor  it = {};
 	int                       id;
 	struct m0_fid             fid;
-	uint8_t                   ht_idx;
+	uint16_t                  ht_idx;
 
 	adom = emap_dom_find(act, &emap_ac->emap_fid, &id);
 	if (adom == NULL || id < 0 || id >= AO_NR - AO_EMAP_FIRST) {
@@ -1404,8 +1405,8 @@ static int emap_prep(struct action *act, struct m0_be_tx_credit *credit)
 		else
 			m0_be_emap_credit(&adom->sad_adata_ht[ht_idx].sad_adata,
 					  M0_BEO_INSERT, 1, credit);
-		m0_be_emap_credit(&adom->sad_adata_ht[ht_idx].sad_adata, M0_BEO_PASTE,
-				  BALLOC_FRAGS_MAX + 1, credit);
+		m0_be_emap_credit(&adom->sad_adata_ht[ht_idx].sad_adata,
+                                 M0_BEO_PASTE, BALLOC_FRAGS_MAX + 1, credit);
 	}
 	m0_mutex_unlock(&beck_builder.b_emaplock[id]);
 	return 0;
@@ -1423,7 +1424,7 @@ static void emap_act(struct action *act, struct m0_be_tx *tx)
 	struct m0_ext             in_ext;
 	int                       id;
 	struct m0_fid             fid;
-	uint8_t			  ht_idx;
+	uint16_t		  ht_idx;
 
 	adom = emap_dom_find(act, &emap_ac->emap_fid, &id);
 	emap_val = emap_ac->emap_val.b_addr;
@@ -1456,11 +1457,11 @@ static void emap_act(struct action *act, struct m0_be_tx *tx)
 		ht_idx = m0_stob_get_hash(&fid);
 
 		rc = rc ? M0_BE_OP_SYNC_RET(op,
-				m0_be_emap_obj_insert(&adom->sad_adata_ht[ht_idx].sad_adata,
-						      tx, &op,
-						      &emap_key->ek_prefix,
-						      AET_HOLE),
-				bo_u.u_emap.e_rc) : 0;
+			    m0_be_emap_obj_insert(&adom->sad_adata_ht[ht_idx].
+                                                        sad_adata, tx, &op,
+                                                 &emap_key->ek_prefix,
+						  AET_HOLE),
+			    bo_u.u_emap.e_rc) : 0;
 
 		M0_SET0(&it);
 		rc = emap_entry_lookup(adom, emap_key->ek_prefix,
@@ -2744,7 +2745,7 @@ static void cob_act(struct action *act, struct m0_be_tx *tx)
 	struct m0_uint128         prefix;
 	int			  id;
 	struct m0_fid 		  fid;
-	uint8_t			  ht_idx;
+	uint16_t		  ht_idx;
 
 	m0_mutex_lock(&beck_builder.b_coblock);
 
@@ -2770,23 +2771,26 @@ static void cob_act(struct action *act, struct m0_be_tx *tx)
 		ht_idx = m0_stob_get_hash(&fid);
 
 		emap_dom_find(&ca->coa_act,
-			      &adom->sad_adata_ht[ht_idx].sad_adata.em_mapping.bb_backlink.bli_fid,
-			      &id);
+			      &adom->sad_adata_ht[ht_idx].sad_adata.em_mapping.
+                                    bb_backlink.bli_fid,
+                             &id);
 		m0_mutex_lock(&beck_builder.b_emaplock[id]);
 		rc = M0_BE_OP_SYNC_RET_WITH(&it.ec_op,
-					    m0_be_emap_lookup(&adom->sad_adata_ht[ht_idx].sad_adata,
-							      &prefix, 0, &it),
-					    bo_u.u_emap.e_rc);
+                                  m0_be_emap_lookup(&adom->sad_adata_ht[ht_idx].
+                                                           sad_adata,
+						     &prefix, 0, &it),
+			           bo_u.u_emap.e_rc);
 		if (rc == 0)
 			m0_be_emap_close(&it);
 		else {
 			rc = M0_BE_OP_SYNC_RET(op,
-					       m0_be_emap_obj_insert(&adom->
-								     sad_adata_ht[ht_idx].sad_adata,
-								     tx, &op,
-								     &prefix,
-								     AET_HOLE),
-					       bo_u.u_emap.e_rc);
+	                        m0_be_emap_obj_insert(&adom->
+                                                           sad_adata_ht[ht_idx].
+                                                           sad_adata,
+		                                      tx, &op,
+						      &prefix,
+						      AET_HOLE),
+                               bo_u.u_emap.e_rc);
 		}
 		m0_mutex_unlock(&beck_builder.b_emaplock[id]);
 	}
